meson: drop default-session-manager, use the first session-manager instead
default-session-manager is only used in the pw-uninstalled.sh developer
script. c25cec230c0619bfb7eb4c56f6325046b3274667 added the ability to
use the system-installed wireplumber but other than that this option is
largely superfluous.

Drop it and pick whichever the first entry to session-manager is. For
the vast use-case of either MS or WP this will just work fine and for
the niche case of building both SMs one just needs to make sure the
order is as desired.

And in the case of no session manager, the config line to start the SM
is now commented out.