pipewire-jack: fix use of `pthread_cancel`
`pthread_cancel` is always undefined behavior when used on arbitrary user code,
and JACK clients are not written with it in mind. Thread cancellation should be
opt-in rather than always-on, and so for the JACK shim it makes the most sense
for the default mode to be joining the process thread (like in JACK and JACK2)
and for hard-killing the thread to be a separate function that the library
consumer must use explicitly.